refactor: use a new naming convention and project folder structure
