+++ v2/rest/tweets/index.ts
@@ -1,0 +1,5 @@
+export * from './ManageTweets';
+export * from './SearchTweets';
+export * from './Timelines';
+export * from './TweetCounts';
+export * from './TweetsLookup';
